requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-knox==4.1.0
django-restframework==0.0.1
django-shopping-cart==0.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
idna==2.10
importlib-metadata==3.3.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==8.0.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
validate-email==1.3
virtualenv==20.2.2
virtualenvwrapper-win==1.2.6
waitress==2.0.0
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
